Fix GLIB GNotification versioning
