+++ .travis.yml
@@ -20,7 +20,7 @@ matrix:
 
 before_install:
   - sudo apt-get update -qq
-  - sudo apt-get install -qq -y gtk+3.0 libgtk-3-dev
+  - sudo apt-get install -qq -y gtk+3.0 libgtk-3-dev libnotify-dev
   - sudo apt-get install -qq -y xvfb
   - "export DISPLAY=:99.0"
   - sudo /usr/bin/Xvfb $DISPLAY 2>1 > /dev/null &
