+++ glib/notifications.go
@@ -8,16 +8,6 @@ package glib
 import "C"
 import "unsafe"
 
-// NotificationPriority is a representation of GLib's GNotificationPriority.
-type NotificationPriority int
-
-const (
-	NOTIFICATION_PRIORITY_NORMAL NotificationPriority = C.G_NOTIFICATION_PRIORITY_NORMAL
-	NOTIFICATION_PRIORITY_LOW    NotificationPriority = C.G_NOTIFICATION_PRIORITY_LOW
-	NOTIFICATION_PRIORITY_HIGH   NotificationPriority = C.G_NOTIFICATION_PRIORITY_HIGH
-	NOTIFICATION_PRIORITY_URGENT NotificationPriority = C.G_NOTIFICATION_PRIORITY_URGENT
-)
-
 // Notification is a representation of GNotification.
 type Notification struct {
 	*Object
@@ -72,11 +62,6 @@ func (v *Notification) SetBody(body string) {
 // SetIcon is a wrapper around g_notification_set_icon
 func (v *Notification) SetIcon(icon *Icon) {
 	C.g_notification_set_icon(v.native(), icon.native())
-}
-
-// SetPriority is a wrapper around g_notification_set_priority().
-func (v *Notification) SetPriority(prio NotificationPriority) {
-	C.g_notification_set_priority(v.native(), C.GNotificationPriority(prio))
 }
 
 // SetDefaultAction is a wrapper around g_notification_set_default_action().
